Added a third solution for Day 9
Main difference to 2nd solution is the addition of a Heap and an AdjacencyMatrix class, which inherits from the builtin list. It performs ~100x faster for the minimum value, but ~ 10x slower for the maximum value. Still no idea why it's so different depending on if it's looking for the shortest or longest distance.